[ECS] Move CodingStandardError to SniffRunner domain
