+++ packages/easy-coding-standard/packages/SniffRunner/DataCollector/SniffMetadataCollector.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Symplify\EasyCodingStandard\Application;
+namespace Symplify\EasyCodingStandard\SniffRunner\DataCollector;
 
 use PHP_CodeSniffer\Sniffs\Sniff;
-use Symplify\EasyCodingStandard\ValueObject\Error\CodingStandardError;
+use Symplify\EasyCodingStandard\SniffRunner\ValueObject\Error\CodingStandardError;
 
 final class SniffMetadataCollector
 {
