+++ packages/easy-coding-standard/packages/SniffRunner/ValueObject/Error/CodingStandardError.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Symplify\EasyCodingStandard\ValueObject\Error;
+namespace Symplify\EasyCodingStandard\SniffRunner\ValueObject\Error;
 
 use Symplify\EasyCodingStandard\Parallel\Contract\Serializable;
 use Symplify\EasyCodingStandard\Parallel\ValueObject\Name;
