+++ packages/easy-coding-standard/src/Contract/Application/FileProcessorInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Symplify\EasyCodingStandard\Contract\Application;
 
-use Symplify\EasyCodingStandard\ValueObject\Error\CodingStandardError;
+use Symplify\EasyCodingStandard\SniffRunner\ValueObject\Error\CodingStandardError;
 use Symplify\EasyCodingStandard\ValueObject\Error\FileDiff;
 use Symplify\SmartFileSystem\SmartFileInfo;
 
@@ -13,7 +13,7 @@ interface FileProcessorInterface
     public function processFileToString(SmartFileInfo $smartFileInfo): string;
 
     /**
-     * @return array<FileDiff|CodingStandardError>
+     * @return array<string, array<FileDiff|CodingStandardError>>
      */
     public function processFile(SmartFileInfo $smartFileInfo): array;
